requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.8.0
django-filter==21.1
djangorestframework==3.12.4
PyJWT==2.1.0
pytz==2022.1
sqlparse==0.4.3
